@@ -1,19 +1,16 @@
 import 'package:flutter/material.dart';
+import './main.dart';
 
 class FourthScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("Fourth Screen"),
+        title: Text("Dashboard Screen"),
       ),
       body: Center(
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          child: Text('Go back!'),
-        ),
+        child: 
+            TheBackButton(),
       ),
     );
   }
